+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `matice` will be documented in this file
 
+## Unreleased
+
+- Allow per-call `only` / `except` overrides on `@translations($locale, $options)` (#27)
+
 ## 1.0.0 - 201X-XX-XX
 
 - initial release